day 12 part 2... not complete
Part two is proving tricky because I tried to do the brute-force
solution, but decided to stop things after 10 minutes of things just
sittng there warming up.

Apparently the solution has to do with Isomorphism and finding the
Least Common Multiple or the Greatest Common Denominator or something
like that. It's something that's going to require a bit of reading on my
part.