Finished day 13
It took me a while because I had the wrong assumption about the intcode
program I was given. I was assuming that it would store the address of
each block in its memory. I only realized this when the end of the
output was something like _o_oo_o_o and then it would crash.

My solution was to create a new Screen class that would store the
positions of all the tiles and reprint them every time they updated.